Makes statistics testable
